+++ warp/Network/Wai/Handler/Warp/Header.hs
@@ -91,6 +91,8 @@ data ResponseHeaderIndex
     | ResServer
     | ResDate
     | ResLastModified
+    | ResTransferEncoding
+    | ResConnection
     deriving (Enum, Bounded)
 
 -- | The size for 'IndexedHeader' for HTTP Response.
@@ -101,8 +103,10 @@ responseKeyIndex :: HeaderName -> Int
 responseKeyIndex hn = case BS.length bs of
     4 | bs == "date" -> fromEnum ResDate
     6 | bs == "server" -> fromEnum ResServer
+    10 | bs == "connection" -> fromEnum ResConnection
     13 | bs == "last-modified" -> fromEnum ResLastModified
     14 | bs == "content-length" -> fromEnum ResContentLength
+    17 | bs == "transfer-encoding" -> fromEnum ResTransferEncoding
     _ -> -1
   where
     bs = foldedCase hn